changed line 6: instead of int, long is now used to store the whole number

+++ src/Bitch.java
@@ -3,7 +3,7 @@ public class Bitch {
         int i = 1;
 
 
-        int x = 1;
+        long x = 1;
         while (i <= 10) {
